Add files via upload
hey, everyone, this is a library management system that contains the list of books available along with the library's name. It has a feature of displaying all the books, lending them to someone, returning the issued book, etc. It also has a feature where the issued book data is stored in a dictionary followed by who issued it. kindly go through it once, it's an awesome project of OOPS in Python, have fun:)